// src/types/PokemonType.ts
export enum PokemonType {
  Bug = "BUG",
  Dark = "DARK",
  Dragon = "DRAGON",
  Electric = "ELECTRIC",
  Fairy = "FAIRY",
  Fighting = "FIGHTING",
  Fire = "FIRE",
  Flying = "FLYING",
  Ghost = "GHOST",
  Grass = "GRASS",
  Ground = "GROUND",
  Ice = "ICE",
  Normal = "NORMAL",
  Poison = "POISON",
  Psychic = "PSYCHIC",
  Rock = "ROCK",
  Steel = "STEEL",
  Water = "WATER",
}